+++ src/main/java/lotto/domain/customer/LottoBundle.java
@@ -1,8 +1,5 @@
 package lotto.domain.customer;
 
-import lotto.domain.processor.LottoRank;
-import lotto.domain.processor.WinningLotto;
-
 import java.util.List;
 import java.util.stream.Collectors;
