+++ src/main/java/lotto/domain/customer/LottoRank.java
@@ -1,4 +1,4 @@
-package lotto.domain.processor;
+package lotto.domain.customer;
 
 import java.util.Arrays;
